+++ runtime/utils.ts
@@ -2,6 +2,17 @@ import { type RequestState, vary } from "../blocks/utils.tsx";
 import { Context } from "../deco.ts";
 import { defaultHeaders } from "../utils/http.ts";
 
+export const sha1 = async (text: string) => {
+  const buffer = await crypto.subtle
+    .digest("SHA-1", new TextEncoder().encode(text));
+
+  const hex = Array.from(new Uint8Array(buffer))
+    .map((b) => b.toString(16).padStart(2, "0"))
+    .join("");
+
+  return hex;
+};
+
 export function numToUint8Array(num: number) {
   const arr = new Uint8Array(8);
   for (let i = 0; i < 8; i++) {
